#Added basic coe for implementing user roles and authorities.

+++ src/main/java/allinone/entities/AbstractEntity.java
@@ -16,24 +16,24 @@ import org.springframework.data.domain.Persistable;
 @Data
 @MappedSuperclass
 public abstract class AbstractEntity implements Persistable<Long> {
-
-	private static final long	serialVersionUID	= 1L;
-	
-	@Id
-	@GeneratedValue ( strategy = GenerationType.AUTO)
-	// , generator = "seq_generator")
-	protected Long	                  id=0L;
-
-	@Version
-	@Column ( name = "version")
-	protected Integer	                  version;
-
-	public Long getId() {
-		return id;
-	}
-
-	public boolean isNew() {
-		return id != null ? true : false;
-	}
-
+    
+    private static final long serialVersionUID = 1L;
+    
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    // , generator = "seq_generator")
+    protected Long            id               = 0L;
+    
+    @Version
+    @Column(name = "version")
+    protected Long            version;
+    
+    public Long getId() {
+        return id;
+    }
+    
+    public boolean isNew() {
+        return id != null ? true : false;
+    }
+    
 }
